+++ OpenTapoWearOs/app/src/main/java/dev/veeso/opentapowearos/view/login_activity/ActivityState.kt
@@ -1,0 +1,6 @@
+package dev.veeso.opentapowearos.view.login_activity
+
+enum class ActivityState {
+    LOGIN_FORM,
+    SIGNING_IN
+}