+++ OpenTapoWearOs/app/src/main/java/dev/veeso/opentapowearos/view/main_activity/ActivityState.kt
@@ -1,0 +1,8 @@
+package dev.veeso.opentapowearos.view.main_activity
+
+enum class ActivityState {
+    DEVICE_LIST,
+    LOADING_DEVICE_LIST,
+    NO_DEVICE_FOUND,
+    NO_LINK;
+}