+++ CHANGELOG.md
@@ -1,7 +1,18 @@
 # Changelog
 
 - [Changelog](#changelog)
+  - [0.2.0](#020)
   - [0.1.0](#010)
+
+## 0.2.0
+
+Released on 30/01/2023
+
+- Device groups:
+  - Added the possibility to create groups
+  - Added the possibility to remove created groups
+- Speeded up device scanner
+- Improved icon graphics
 
 ## 0.1.0